refactor: merge mediaBackup actions and reducer
